cannot delete last .tex file
